Task 12 Extracting a Method
Used Extract Method refactor operation on only the for loop rather than
the string and for loop. This is because after running JUnit test with
all tests valid it showed that only the for loop was (0.000s) while the
for loop and string was (0.002s).
JUnit test shows that extracting method refactor operation on only the
for loop was the best option.